+++ src/full/Agda/Interaction/Options/Default.hs
@@ -45,6 +45,7 @@ defaultOptions = Options
   , optTransliterate         = False
   , optDiagnosticsColour     = AutoColour
   , optMdOnlyAgdaBlocks      = False
+  , optParallelChecking      = Sequential
   }
 
 defaultPragmaOptions :: PragmaOptions