+++ CHANGELOG.md
@@ -11,6 +11,11 @@ Installation
 Pragmas and options
 -------------------
 
+* New option `--parallel` (`-j` for short, as in `make`) lets Agda
+  type-check imported modules in parallel. For large libraries, this can
+  result in a speedup of ~3x, at the cost of roughly doubling the
+  maximum memory usage.
+
 * New option `--print-options` to print a simple list of all options.
   This list can e.g. be used to implement bash completion.